+++ packages/form-store/package.json
@@ -22,15 +22,15 @@
   "scripts": {
     "build": "tsc --project tsconfig-build.json",
     "clean": "rimraf ./build",
-    "test": "jest",
+    "test": "jest --silent",
     "verify": "tsc --noEmit"
   },
   "dependencies": {
     "@rixio/atom": "^0.3.9"
   },
   "devDependencies": {
-    "@hapi/joi": "17.1.1",
-    "@types/hapi__joi": "17.1.6",
+    "joi": "^17.4.0",
+    "@types/joi": "^17.2.0",
     "@types/jest": "26.0.0",
     "jest": "26.4.2",
     "rxjs": "6.6.3",
