+++ packages/form-store/src/domain.ts
@@ -1,16 +1,12 @@
-import { Observable } from "rxjs"
+import type { Observable } from "rxjs"
 
-export type ValidationStatus = "validating" | "error" | "success"
+export type ValidationResultValidating = { status: "validating" }
+export const validationResultValidating: ValidationResultValidating = { status: "validating" }
 
-export interface ValidationResultValidating {
-	status: "validating"
-}
+export type ValidationResultSuccess = { status: "success" }
+export const validationResultSuccess: ValidationResultSuccess = { status: "success" }
 
-export interface ValidationResultSuccess {
-	status: "success"
-}
-
-export interface ValidationResultError<T> {
+export type ValidationResultError<T> = {
 	status: "error"
 	error: string
 	children: {
@@ -19,7 +15,7 @@ export interface ValidationResultError<T> {
 }
 
 export type ValidationResult<T> = ValidationResultSuccess | ValidationResultValidating | ValidationResultError<T>
-
+export type ValidationStatus = ValidationResult<any>["status"]
 export type Validate<T> = (
 	value: T
 ) => ValidationResult<T> | PromiseLike<ValidationResult<T>> | Observable<ValidationResult<T>>
