+++ packages/form-store/src/is-submit-disabled.ts
@@ -1,6 +1,6 @@
 import { combineLatest, Observable } from "rxjs"
 import { map } from "rxjs/operators"
-import { FormStore } from "./index"
+import type { FormStore } from "./index"
 
 export function isSubmitDisabled(form: FormStore<any>, displayErrors: Observable<boolean>) {
 	return combineLatest([form.canSubmit$, displayErrors]).pipe(
