+++ packages/form-store/src/utils/create-validation-result.ts
@@ -1,7 +1,7 @@
 import { concat, Observable, of } from "rxjs"
 import { fromPromise } from "rxjs/internal-compatibility"
 import { mergeMap, scan, shareReplay } from "rxjs/operators"
-import { Validate, ValidationResult, ValidationResultValidating } from "../domain"
+import { Validate, ValidationResult, validationResultValidating } from "../domain"
 
 export function createValidationResult<T>(
 	value: Observable<T>,
@@ -10,15 +10,12 @@ export function createValidationResult<T>(
 	const simplified = simplify(validate)
 	return value.pipe(
 		mergeMap(simplified),
-		scan<ValidationResult<T>>(
-			(prev, next) => {
-				if (next.status === "validating" && prev.status === "error") {
-					return prev
-				}
-				return next
-			},
-			{ status: "validating" } as ValidationResultValidating
-		),
+		scan<ValidationResult<T>>((prev, next) => {
+			if (next.status === "validating" && prev.status === "error") {
+				return prev
+			}
+			return next
+		}, validationResultValidating),
 		shareReplay(1)
 	)
 }
@@ -28,10 +25,10 @@ function simplify<T>(validate: Validate<T>): (value: T) => Observable<Validation
 		const vr = validate(value)
 		if (vr instanceof Observable) {
 			return vr
-		} else if ("then" in vr) {
-			return concat(of({ status: "validating" } as ValidationResultValidating), fromPromise(vr))
-		} else {
-			return of(vr)
 		}
+		if ("then" in vr) {
+			return concat(of(validationResultValidating), fromPromise(vr))
+		}
+		return of(vr)
 	}
 }
