+++ packages/form-store/src/utils/validate-joi.ts
@@ -1,5 +1,5 @@
-import type { ObjectSchema, ValidationOptions } from "@hapi/joi"
-import type { ValidationResult, ValidationResultError, ValidationResultSuccess } from "../domain"
+import type { ObjectSchema, ValidationOptions } from "joi"
+import { ValidationResult, ValidationResultError, validationResultSuccess } from "../domain"
 
 export function validateJoi<T>(
 	schema: ObjectSchema,
@@ -20,13 +20,17 @@ export function validateJoi<T>(
 			vr.error.details.forEach(err => {
 				let current: ValidationResultError<any> = result
 				err.path.forEach(path => {
-					const next: ValidationResultError<any> = { status: "error", error: err.message, children: {} }
+					const next: ValidationResultError<any> = {
+						status: "error",
+						error: err.message,
+						children: {},
+					}
 					current.children[path] = next
 					current = next
 				})
 			})
 			return result as ValidationResult<T>
 		}
-		return { status: "success" } as ValidationResultSuccess
+		return validationResultSuccess
 	}
 }
